pipeline/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
efficientnet_pytorch
retinaface_pytorch
opencv-python
numpy
Pillow
tqdm
scikit-learn
pandas
gdown
pretrainedmodels
pytest
pretrainedmodels
pytest
